+++ sq/cabal.nix
@@ -2,11 +2,11 @@
 , attoparsec-aeson, attoparsec-iso8601, base, binary, bytestring
 , clock, containers, contravariant, criterion, deepseq, df1, di
 , di-core, di-df1, direct-sqlite, directory, exceptions, filepath
-, foldl, hedgehog, integer-logarithms, lib, profunctors, ref-tf
-, resource-pool, resourcet, resourcet-extra, safe-exceptions
-, scientific, sop-core, stm, streaming, tasty, tasty-hedgehog
-, tasty-hunit, template-haskell, text, time, transformers
-, uuid-types
+, foldl, generics-sop, hedgehog, integer-logarithms, lib
+, profunctors, ref-tf, resource-pool, resourcet, resourcet-extra
+, safe-exceptions, scientific, sop-core, stm, streaming, tasty
+, tasty-hedgehog, tasty-hunit, template-haskell, text, time
+, transformers, uuid-types
 }:
 mkDerivation {
   pname = "sq";
@@ -16,9 +16,9 @@ mkDerivation {
     adjunctions aeson attoparsec attoparsec-aeson attoparsec-iso8601
     base binary bytestring clock containers contravariant deepseq
     di-core di-df1 direct-sqlite directory exceptions filepath foldl
-    integer-logarithms profunctors ref-tf resource-pool resourcet
-    resourcet-extra safe-exceptions scientific sop-core stm streaming
-    template-haskell text time transformers uuid-types
+    generics-sop integer-logarithms profunctors ref-tf resource-pool
+    resourcet resourcet-extra safe-exceptions scientific sop-core stm
+    streaming template-haskell text time transformers uuid-types
   ];
   testHaskellDepends = [
     aeson async base binary bytestring df1 di di-core hedgehog ref-tf